- Добавил класс для работы с консолью ConsoleManager - Удалил класс ResultTable - Мелкие изменения
