+++ GeniusIdiotConsoleApp/UserStorage.cs
@@ -1,19 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
-
-/*
-    В качестве бонусного задания, можно немного разделить логику работы программы.
-    Выделить класс для работы с файловой системой и класс для работы с консолью(опционально)
-*/
+using System.IO;
+using System.Linq;
 
 namespace GeniusIdiotConsoleApp
 {
+    /// <summary>
+    /// Статический класс для данных пользователей
+    /// </summary>
     public static class UserStorage
     {
         private static string filePath { get; } = @".\TestResults\TestResults.txt";
         public static IEnumerable<User> GetUsersResults()
         {
-            var lines = FileManager.GetContent(filePath).Split('\n');
+            var lines = FileManager.GetContent(filePath).Split(new string[] {"\n"}, StringSplitOptions.RemoveEmptyEntries);
             foreach (var userResult in lines)
             {
                 var userInfo = userResult.Split(new string[] { "|||||" }, StringSplitOptions.RemoveEmptyEntries);
